requirements.txt
flask-bootstrap==3.3.7.1
flask_sqlalchemy==3.1.1
Flask==3.0.3
SQLAlchemy==2.0.25
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Bootstrap_Flask==2.3.3
Flask_WTF==1.2.1
WTForms==3.0.1
Flask_CKEditor==0.5.1
